requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
hydra-core
liger-kernel
numpy<2.0.0
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict>=0.8.0,<=0.10.0,!=0.9.0
torchdata
transformers
# vllm==0.8.4
wandb
packaging>=20.0
uvicorn
fastapi
latex2sympy2_extended
math_verify
tensorboard


######################################
# 模型训练相关
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
deepspeed>=0.10.0
tqdm>=4.66.0
matplotlib>=3.7.0
tensorboard>=2.15.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 文本处理
googletrans==4.0.0rc1

# 通信和配置
pyzmq>=25.0.0
pyyaml>=6.0

# PDF处理
grobid-client==0.8.8
grobid-client-python==0.1.2
pymupdf==1.26.7

# HTTP请求
requests==2.32.5

bottle
math-verify
safetensors
sentencepiece
tornado

